+++ workspaces/jfrog-artifactory/plugins/jfrog-artifactory/dev/index.tsx
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2024 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import React from 'react';
+
+import { Entity } from '@backstage/catalog-model';
+import { createDevApp } from '@backstage/dev-utils';
+import { EntityProvider } from '@backstage/plugin-catalog-react';
+
+import { getAllThemes } from '@redhat-developer/red-hat-developer-hub-theme';
+
+import { JfrogArtifactoryPage, jfrogArtifactoryPlugin } from '../src/plugin';
+
+const mockEntity: Entity = {
+  apiVersion: 'backstage.io/v1alpha1',
+  kind: 'Component',
+  metadata: {
+    name: 'backstage',
+    description: 'backstage.io',
+    annotations: {
+      'jfrog-artifactory/image-name': 'backstage',
+    },
+  },
+  spec: {
+    lifecycle: 'production',
+    type: 'service',
+    owner: 'user:guest',
+  },
+};
+
+createDevApp()
+  .registerPlugin(jfrogArtifactoryPlugin)
+  .addThemes(getAllThemes())
+  .addPage({
+    element: (
+      <EntityProvider entity={mockEntity}>
+        <JfrogArtifactoryPage />
+      </EntityProvider>
+    ),
+    title: 'Root Page',
+    path: '/artifactory',
+  })
+  .render();