+++ workspaces/redhat-argocd/packages/app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # app
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [a1313aa]
+  - @backstage-community/plugin-redhat-argocd@1.8.6
+
 ## 0.0.4
 
 ### Patch Changes